Switch back to semantic versioning
Because goreleaser requires it: https://goreleaser.com/semver/

Originally moved away from it because this package is only meant for
personal usage, so using a date to increment the version is quicker than
figuring out what changed to keep the semantic versioning correct.

Go to 2.0.0 since we introduced breaking changes compared to upstream.